adding props to profitability table
